requirements.txt
absl-py==2.0.0
altair==4.0.0
altgraph==0.17.3
amqp==5.1.1
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
auto-py-to-exe==2.34.0
beautifulsoup4==4.12.2
billiard==4.1.0
bleach==6.0.0
blinker==1.8.2
boto3==1.28.35
botocore==1.31.35
bottle==0.12.25
bottle-websocket==0.2.9
cachetools==5.3.2
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
comm==0.1.4
contourpy==1.1.1
crispy-bootstrap4==2023.1
crispy-bootstrap5==2023.10
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
distro==1.9.0
Django==4.2.4
django-auto-logout==0.5.1
django-background-tasks==1.2.5
django-bootstrap4==23.2
django-ckeditor==6.7.0
django-ckeditor-5==0.2.11
django-compat==1.0.15
django-cors-headers==4.4.0
django-crispy-forms==2.1
django-db-logger==0.1.12
django-editorjs==0.2.1
django-jazzmin==2.6.0
django-js-asset==2.1.0
django-paypal==2.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.20
django-storages==1.13.2
django-taggit==3.0.0
django4-background-tasks==1.2.9
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Eel==0.16.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==2.0.1
flatbuffers==23.5.26
fonttools==4.43.1
future==0.18.3
gast==0.5.4
gevent==22.10.2
gevent-websocket==0.10.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.59.2
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.5
httpx==0.27.0
idna==3.4
importlib_metadata==8.0.0
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.5.0
jupyter_core==5.5.0
keras==2.14.0
kiwisolver==1.4.5
kombu==5.3.1
libclang==16.0.6
macholib==1.16.2
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.2.0
nest-asyncio==1.5.8
nose==1.3.7
numpy==1.26.1
oauthlib==3.2.2
openai==1.35.10
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.1.2
parso==0.8.3
paypal==1.2.5
pdfkit==1.0.0
pexpect==4.8.0
Pillow==10.0.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.6
psycopg2==2.9.7
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.16.1
pyinstaller==5.10.1
pyinstaller-hooks-contrib==2023.2
PyJWT==2.8.0
Pympler==1.1
pyparsing==3.0.9
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytube3==9.6.4
pytz==2023.3.post1
pyzmq==25.1.1
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
requests-oauthlib==1.3.1
revChatGPT==1.1.3
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
s3transfer==0.6.2
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
semver==3.0.2
shortuuid==1.0.11
slack_sdk==3.32.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
streamlit==1.15.2
stripe==8.1.0
tenacity==8.5.0
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
threadpoolctl==3.2.0
tiktoken==0.7.0
toml==0.10.2
toolz==0.12.1
tornado==6.3.3
tqdm==4.66.4
traitlets==5.13.0
types-python-dateutil==2.8.19.14
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.16
validators==0.30.0
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==3.0.1
whichcraft==0.6.1
whitenoise==6.5.0
wrapt==1.14.1
zipp==3.19.2
zope.event==4.6
zope.interface==6.0
